5.14. Модуляция.
