+++ index.html
@@ -233,5 +233,7 @@
   </template>
   <script src="js/functions.js"></script>
   <script src="js/main.js"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/util.js" type="module"></script>
 </body>
 </html>
